+++ forge/src/main/java/com/bookkeepersmc/notebook/forge/NotebookForge.java
@@ -1,5 +1,6 @@
-package com.bookkeepersmc.notebook;
+package com.bookkeepersmc.notebook.forge;
 
+import com.bookkeepersmc.notebook.NotebookCommon;
 import net.minecraftforge.fml.common.Mod;
 
 @Mod(NotebookCommon.MOD_ID)