+++ CHANGELOG.md
@@ -1,2 +1,2 @@
 ### Additions
-- BlockEntityBuilder
+- Neo/Forge Datagen Helpers